redirectURL

+++ src/app/pages/database/database.component.ts
@@ -3,12 +3,13 @@ import {Phrase} from '../../interfaces/phrase';
 import {PhraseService} from '../../services/phrase.service';
 import {Router, RouterLink} from '@angular/router';
 import {AccordionComponent} from '../../accordion/accordion.component';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-database',
   imports: [
     RouterLink,
-    AccordionComponent
+    AccordionComponent,
   ],
   templateUrl: './database.component.html',
   styleUrl: './database.component.css'
@@ -16,12 +17,18 @@ import {AccordionComponent} from '../../accordion/accordion.component';
 export class DatabaseComponent implements OnInit {
 
   phrases:Phrase[] = [];
+  needsToReviewsPhrases:Phrase[] = [];
+  hiddenPhrases:Phrase[] = [];
+  newPhrases:Phrase[] = [];
 
   constructor(private phraseService:PhraseService,private router: Router) {
   }
 
   ngOnInit() {
     this.phrases=this.phraseService.getPhrases();
+    this.needsToReviewsPhrases=this.phrases.filter(item => item.needToReview);
+    this.hiddenPhrases=this.phrases.filter(item => item.hide);
+    this.newPhrases=this.phrases.filter(item => item.needToReview && !item.hide);
   }
 
   toggleAccordion(index: number) {
@@ -29,7 +36,19 @@ export class DatabaseComponent implements OnInit {
   }
 
   deletePhrase($event: string) {
-    this.phraseService.deletePhrase($event)
-    this.phrases = this.phrases.filter(phrase => phrase.id !== $event);
+    Swal.fire({
+      icon: "error",
+      theme:'dark',
+      text: "Do you want to Delete This Phrase?",
+      showCancelButton: true,
+      confirmButtonText: "Yes",
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        this.phraseService.deletePhrase($event)
+        this.phrases = this.phrases.filter(phrase => phrase.id !== $event);
+      }
+    });
+
   }
 }
